tidy(ui): rename builder dnd file
